Migrate Box.jsx to tsx
